Add files via upload
Deep Ensemble Models (DF/DNDT/DNDF/DSVM)
Run the main.py file to execute the code and then choose the model number ranging from 1 to 4 to execute the specific deep ensemble model. Continue execution to explain the model by choosing y and n in the console. LIME and SHAP works for all the 4 models but LEAF is executable only for DF and DSVM. Not yet compatible with DNDT and DNDF. Works only with lower versions of Python (3.8 or lower)